# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateRegionRequest:
    properties:
      coord_x:
        type: integer
      coord_y:
        type: integer
      max_capacity:
        type: integer
      region_id:
        type: string
      region_name:
        type: string
    required:
    - region_id
    - region_name
    type: object
  dto.CreateTripLogBRequest:
    properties:
      destination_1:
        type: string
      destination_2:
        type: string
      destination_3:
        type: string
      end_time:
        description: RFC3339 string
        type: string
      start_time:
        description: RFC3339 string
        type: string
      status:
        description: '"운행중" or "비운행중"'
        type: string
      vehicle_id:
        type: string
    required:
    - vehicle_id
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  dto.RegionResponse:
    properties:
      coord_x:
        type: integer
      coord_y:
        type: integer
      current_capacity:
        type: integer
      is_full:
        type: boolean
      max_capacity:
        type: integer
      region_id:
        type: string
      region_name:
        type: string
      saturated_at:
        type: string
    type: object
  dto.TripLogBResponse:
    properties:
      destination_1:
        type: string
      destination_2:
        type: string
      destination_3:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
      trip_id:
        type: integer
      vehicle_id:
        type: string
    type: object
...
  dto.UpdateRegionRequest:
    properties:
      coord_x:
        type: integer
      coord_y:
        type: integer
      current_capacity:
        type: integer
      is_full:
        type: boolean
      max_capacity:
        type: integer
      region_name:
        type: string
      saturated_at:
        type: string
    required:
    - region_name
    type: object
  dto.UpdateTripLogBRequest:
    properties:
      destination_1:
        type: string
      destination_2:
        type: string
      destination_3:
        type: string
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 패키지 운송 시스템 API 문서입니다.
  title: Go API Server
  version: "1.0"
paths:
...
  /api/region:
    get:
      description: 모든 지역 정보를 반환합니다.
      parameters:
      - description: '정렬 필드 (예: -registered_at는 최신순, region_id 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RegionResponse'
            type: array
      summary: 모든 지역 조회
      tags:
      - region
    post:
      consumes:
      - application/json
      description: 새로운 지역을 생성합니다.
      parameters:
      - description: 지역 정보
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRegionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 지역 생성
      tags:
      - region
  /api/region/{id}:
    delete:
      description: 지역 ID로 지역 정보를 삭제합니다.
      parameters:
      - description: 지역 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 지역 삭제
      tags:
      - region
    get:
      description: 지역 ID로 지역 정보를 조회합니다.
      parameters:
      - description: 지역 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 지역 단건 조회
      tags:
      - region
    put:
      consumes:
      - application/json
      description: 지역 ID로 지역 정보를 수정합니다.
      parameters:
      - description: 지역 ID
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 지역 정보
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRegionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 지역 정보 수정
      tags:
      - region
  /api/region/search:
    get:
      description: 쿼리 파라미터로 지역을 검색합니다.
      parameters:
      - description: 지역 ID
        in: query
        name: region_id
        type: string
      - description: 지역명
        in: query
        name: region_name
        type: string
      - description: X 좌표
        in: query
        name: coord_x
        type: integer
      - description: Y 좌표
        in: query
        name: coord_y
        type: integer
      - description: 최대 용량
        in: query
        name: max_capacity
        type: integer
      - description: 현재 용량
        in: query
        name: current_capacity
        type: integer
      - description: 포화 여부
        in: query
        name: is_full
        type: boolean
      - description: 포화 시각(RFC3339)
        in: query
        name: saturated_at
        type: string
      - description: '정렬 필드 (예: -region_id, -max_capacity, -saturated_at 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RegionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 지역 검색
      tags:
      - region
  /api/trip-log-b:
    get:
      description: 모든 B차 운행 로그 정보를 반환합니다.
      parameters:
      - description: '정렬 필드 (예: -trip_id, -start_time 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TripLogBResponse'
            type: array
      summary: 모든 B차 운행 로그 조회
      tags:
      - trip_log_b
    post:
      consumes:
      - application/json
      description: 새로운 B차 운행 로그를 생성합니다.
      parameters:
      - description: B차 운행 로그 정보
        in: body
        name: trip_log_b
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTripLogBRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TripLogBResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: B차 운행 로그 생성
      tags:
      - trip_log_b
  /api/trip-log-b/{id}:
    delete:
      description: trip_id로 B차 운행 로그를 삭제합니다.
      parameters:
      - description: B차 운행 로그 trip_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: B차 운행 로그 삭제
      tags:
      - trip_log_b
    get:
      description: trip_id로 B차 운행 로그를 조회합니다.
      parameters:
      - description: B차 운행 로그 trip_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TripLogBResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: B차 운행 로그 단건 조회
      tags:
      - trip_log_b
    put:
      consumes:
      - application/json
      description: trip_id로 B차 운행 로그 정보를 수정합니다.
      parameters:
      - description: B차 운행 로그 trip_id
        in: path
        name: id
        required: true
        type: integer
      - description: 수정할 B차 운행 로그 정보
        in: body
        name: trip_log_b
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTripLogBRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TripLogBResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: B차 운행 로그 정보 수정
      tags:
      - trip_log_b
  /api/trip-log-b/search:
    get:
      description: 쿼리 파라미터로 B차 운행 로그를 검색합니다.
      parameters:
      - description: trip_id
        in: query
        name: trip_id
        type: integer
      - description: 차량 ID
        in: query
        name: vehicle_id
        type: string
      - description: 상태
        in: query
        name: status
        type: string
      - description: 목적지1
        in: query
        name: destination_1
        type: string
      - description: 목적지2
        in: query
        name: destination_2
        type: string
      - description: 목적지3
        in: query
        name: destination_3
        type: string
      - description: '정렬 필드 (예: -trip_id, -start_time 등)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TripLogBResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: B차 운행 로그 검색
      tags:
      - trip_log_b
...
swagger: "2.0"